ref(eap): Extract attribute validation helpers into dedicated module
Move serialize_type, _check_attributes_by_type, and _check_attributes_exist
from organization_trace_item_attributes into search/eap/attribute_validation
so they can be reused by other modules.
